Refactor codes and added error handling in the absence of the Internet
