+++ src/pages/PassingTestPage/PassingTestFooter/index.js
@@ -1,0 +1,1 @@
+export { default } from './PassingTestFooter';
